# tracknova/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TrackNova public transport tracking API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type RouteType = typeof RouteType[keyof typeof RouteType];


export const RouteType = {
  bus: 'bus',
  metro: 'metro',
  'suburban-rail': 'suburban-rail',
} as const;

export interface Route {
  id: number;
  name: string;
  number: string;
  type: RouteType;
  color: string;
  startStop: string;
  endStop: string;
  totalStops: number;
  frequency: string;
  isActive?: boolean;
}

export type StopType = typeof StopType[keyof typeof StopType];


export const StopType = {
  bus: 'bus',
  metro: 'metro',
  'suburban-rail': 'suburban-rail',
} as const;

export interface Stop {
  id: number;
  name: string;
  lat: number;
  lng: number;
  type: StopType;
  /** @nullable */
  address?: string | null;
  routeIds?: number[];
}

export interface RouteDetail {
  id: number;
  name: string;
  number: string;
  type: string;
  color: string;
  startStop?: string;
  endStop?: string;
  frequency?: string;
  isActive?: boolean;
  stops: Stop[];
}

export interface Arrival {
  routeId: number;
  routeNumber: string;
  routeName: string;
  minutesAway: number;
  vehicleId: string;
  direction?: string;
}

export interface StopDetail {
  id: number;
  name: string;
  lat: number;
  lng: number;
  type: string;
  /** @nullable */
  address?: string | null;
  arrivals: Arrival[];
}

export type VehicleOccupancy = typeof VehicleOccupancy[keyof typeof VehicleOccupancy];


export const VehicleOccupancy = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface Vehicle {
  id: string;
  routeId: number;
  lat: number;
  lng: number;
  speed: number;
  direction: number;
  occupancy: VehicleOccupancy;
  lastUpdated: string;
}

export type DashboardSummaryRoutesByType = {
  bus?: number;
  metro?: number;
  'suburban-rail'?: number;
};

export interface DashboardSummary {
  totalRoutes: number;
  activeVehicles: number;
  totalStops: number;
  onTimePercentage: number;
  routesByType: DashboardSummaryRoutesByType;
}

export type FavoriteItemType = typeof FavoriteItemType[keyof typeof FavoriteItemType];


export const FavoriteItemType = {
  route: 'route',
  stop: 'stop',
} as const;

export interface Favorite {
  id: number;
  userId: string;
  itemId: number;
  itemType: FavoriteItemType;
  itemName: string;
}

export interface FavoritesResponse {
  routes: Favorite[];
  stops: Favorite[];
}

export type FavoriteInputItemType = typeof FavoriteInputItemType[keyof typeof FavoriteInputItemType];


export const FavoriteInputItemType = {
  route: 'route',
  stop: 'stop',
} as const;

export interface FavoriteInput {
  itemId: number;
  itemType: FavoriteInputItemType;
  itemName: string;
}

export type ListRoutesParams = {
type?: ListRoutesType;
search?: string;
};

export type ListRoutesType = typeof ListRoutesType[keyof typeof ListRoutesType];


export const ListRoutesType = {
  bus: 'bus',
  metro: 'metro',
  'suburban-rail': 'suburban-rail',
} as const;

export type ListStopsParams = {
lat?: number;
lng?: number;
radius?: number;
};

export type ListVehiclesParams = {
routeId?: number;
};