# tracknova/lib/api-zod/src/generated/types/favoriteInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TrackNova public transport tracking API
 * OpenAPI spec version: 0.1.0
 */
import type { FavoriteInputItemType } from './favoriteInputItemType';

export interface FavoriteInput {
  itemId: number;
  itemType: FavoriteInputItemType;
  itemName: string;
}