# tracknova/lib/api-zod/src/generated/types/favoritesResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TrackNova public transport tracking API
 * OpenAPI spec version: 0.1.0
 */
import type { Favorite } from './favorite';

export interface FavoritesResponse {
  routes: Favorite[];
  stops: Favorite[];
}