# tracknova/lib/api-zod/src/generated/types/route.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TrackNova public transport tracking API
 * OpenAPI spec version: 0.1.0
 */
import type { RouteType } from './routeType';

export interface Route {
  id: number;
  name: string;
  number: string;
  type: RouteType;
  color: string;
  startStop: string;
  endStop: string;
  totalStops: number;
  frequency: string;
  isActive?: boolean;
}