# tracknova/lib/api-zod/src/generated/types/routeDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TrackNova public transport tracking API
 * OpenAPI spec version: 0.1.0
 */
import type { Stop } from './stop';

export interface RouteDetail {
  id: number;
  name: string;
  number: string;
  type: string;
  color: string;
  startStop?: string;
  endStop?: string;
  frequency?: string;
  isActive?: boolean;
  stops: Stop[];
}