# tracknova/lib/api-zod/src/generated/types/stop.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TrackNova public transport tracking API
 * OpenAPI spec version: 0.1.0
 */
import type { StopType } from './stopType';

export interface Stop {
  id: number;
  name: string;
  lat: number;
  lng: number;
  type: StopType;
  /** @nullable */
  address?: string | null;
  routeIds?: number[];
}