# tracknova/lib/api-zod/src/generated/types/stopDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TrackNova public transport tracking API
 * OpenAPI spec version: 0.1.0
 */
import type { Arrival } from './arrival';

export interface StopDetail {
  id: number;
  name: string;
  lat: number;
  lng: number;
  type: string;
  /** @nullable */
  address?: string | null;
  arrivals: Arrival[];
}