# tracknova/lib/api-zod/src/generated/types/vehicle.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TrackNova public transport tracking API
 * OpenAPI spec version: 0.1.0
 */
import type { VehicleOccupancy } from './vehicleOccupancy';

export interface Vehicle {
  id: string;
  routeId: number;
  lat: number;
  lng: number;
  speed: number;
  direction: number;
  occupancy: VehicleOccupancy;
  lastUpdated: string;
}